+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - PR #1091 Add `indicator=` and `suffixes=` keywords to merge
 - PR #1107 Add unsupported keywords to Series.fillna
 - PR #1136 Removed `gdf_concat` 
+- PR #1148 Add cudf.sqrt for dataframes and Series
 
 ## Improvements